Configure for postgresql

+++ backend/src/main/resources/db/migration/V1__initial_structure.sql
@@ -1,15 +1,15 @@
 CREATE TABLE shop
 (
     id                 uuid primary key,
-    name               varchar not null,
-    owner_name         varchar not null,
-    street             varchar not null,
-    zip_code           varchar not null,
-    city               varchar not null,
-    address_supplement varchar not null,
-    enabled            boolean not null,
-    latitude           double  not null,
-    longitude          double  not null
+    name               varchar          not null,
+    owner_name         varchar          not null,
+    street             varchar          not null,
+    zip_code           varchar          not null,
+    city               varchar          not null,
+    address_supplement varchar          not null,
+    enabled            boolean          not null,
+    latitude           double precision not null,
+    longitude          double precision not null
 );
 
 CREATE TABLE admin
